fix(icon): rename clinical_notesからclinical-notesに変更
