+++ packages/wiz-ui-next/src/components/icons/clinical-notes.vue
